Top-level nav now reflects nav selection
Experimented with methods to reflect navigation selection to the top-level "tab" selection including a custom element that listened for router events and one that listened for an event from each page's render in order to catch any possible display of a page.  In the end, the only way I found that worked reliably was to have app.ts, which displays the links, listen for the custom bubbled event and manually set the classes on each link based on the contents of the event.  Ugly, but simple, fast, reliable, and functional.