Added content to post element page, fixed post id bug
The post tile Read More button wasn't sending the post object but the "readMore" string again in the detail object, so the router URL param was "undefined" instead of the id number.  This is now fixed.

Added a basic stub for the post content to the post element page to show that it's working.  It uses the same data-in-typescript file hack that the posts page does.