SPA routing now works in test browser
In order to support deep linking, bookmarking, etc. in an SPA app with a router like Vaadin, you need to configure the server to handle it by setting the "app index" value to the "index.html" file that forms the root of the app URL.  This change does so for the local test browser.