Use installed Shoelace, fixing race condition with Shadow DOM
When using the CDN Shoelace theme and module, the first load would usually show the content above the navbar because Shoelace returned too late to style it correctly.  Lit didn't correctly position the content in the <slot/> because I had disabled Shadow DOM with createRenderRoot() {this} in order to get the root CSS to affect the App component.

Instead, I figured out how to get the CSS to work: use the styles.js version of the theme instead of the css.  I had to create a .d.ts of the style.js file, but otherwise it worked.

Now the Shoelace module and CSS are bundled and I can support widgets one-by-one in my components via import statements.

I only needed to build once, to affect the asset copy.  The normal tsc -watch web dev server still works.

createRenderRoot() has been removed so that the lit Shadow DOM works as expected once more.