Posts can now have inline SVG card images
Post data now includes 4 new properties related to the card tile: hdrWGL, hdrSVG, hdrImg, hdrAlt.  The first three are inline WebGL, inline SVG, and Image URL content for the image slot in the card, with hdrAlt accessibility text.  Only one is used, in that priority order.

Currently examples are implemented for hdrSVG and hdrImg.

Added error SVG image display for when none of the above is specified.

Added lit-html as package dependency.