Clearing to a color working on WebGL canvas in post tile
The post-tile has to wait for the firstUpdated() event, then check to see if the "post" field has been propagated and includes webGL info, and then waits on the the updateComplete Promise in order to finally initialize webGL against the canvas.  Otherwise, it won't be able to find the element in the DOM.

Also, WebGL now takes the shadowDOM element from Post-Tile in order to help find the canvas element.

There is a lot more data in both the WebGLScene and WebGLObject classes.  It tries to create basic primitives if requested, but they don't render yet.  It does clear to the requested values including color, however.

Removed unnecessary "webglscenedefaults.ts" as the defaults live in webglscene.ts next to the class itself, as an exported object.