Replaced Lint and unwelcome commit-time format step
When saving a source file in VS Code, the typescript formatter applies desired formatting.  When committing in GitKraken, it applies pre-commit hooks via husky including a lint script which ran both eslint --fix and prettier, which was still installed as a devDependency apparently.  This completely changed the formatting of the committed files and caused a potential diff churn.

The prettier component and step has been removed.

The linter has been updated to align with the VS Code formatting rules.

All files have been reformatted to the VS Code format, without any functionality changes.