WebGL supports multiple elements, routing away, resizing and scrolling
Changed from a canvas per WebGL use to a single canvas filling the page, as the first child of the body.

Added WebGLElement class and file to handle rendering and state for a single DOM WebGL element, whereas the WebGL file/class handles the main WebGL state, clearing the entire canvas, and managing/iterating the elements to render.

Post_Item now has a min-height to preserve its shape in the place of the missing canvas.

All WebGL-using LitElements now have an onAfterLeave() method which calls the WebGL.onNavigateAway method to clear all elements so that they don't try to render on the new page.  This is necessary instead of subscribing to the raw vaadin window navigate event directly because that one also triggers when navigating to the same page, which clears the WebGL elements but doesn't replace them.  This currently applies to Post_Item and Post_List.

Turned WebGL class into a singleton.  It holds the reference to the gl context and canvas, the list of elements, and information required to update the rendering (like the most recent render request ID and whether any of the elements are animated, requiring constantly refreshing renders).  When it renders, it resizes the canvas to the display size, updates its transform based on vertical scrolling, then clears the entire screen.  Then it iterates all elements to render them, passing in the time in MS.

The WebGLObject now takes the selector of the DOM element (.class or #id) to which it is attached.