WebGL supports basic object animations (transform and color)
The data file object format now has the xform take an array of xforms, with a new key object per xform, which contain a time value and an optional blending mode to move into that key (linear, discrete/pop, or sine).  Pos and RotAxis are vectors instead of number arrays.  Scale is now a vector instead of scalar.  The object has two new members as well: an array of color objects each containing a color vector and alpha number, as well as a matching key object to the transform; and a new anim object containing a default blend mode for all keys and well as the loop mode (repeat or pingpong) for the whole object.

New webglanim.ts file to hold interfaces and a class supporting animation.

Updated the tex.fs shader to use a uniform color value and multiply it against the diffuse texture.

The WebGL class now properly only renders when required, which means when requested by a resize or scroll or navigation event, or when one of the elements in an object is actually animated (ie, has a key with a time value > 0).

Most of the animation meat is in webglobject.ts.  There is unfortunately poor code reuse between the xform and color methods.  It is probably worth revisiting after checkin for maintainability purposes.