Renamed WebGLElement to WebGLViewport
WebGLElement is now WebGLViewport, to reflect what it actually is and does.  It's a runtime object reflecting a single viewport on the WebGL Canvas.  There can be multiple on the screen.  There will be one Scene per Viewport, but the Scene will come from data (offline) while the Viewport is runtime.