JSON WebGL data now correctly interpreted
The data structure now has a field which can take JSON data representing the WebGL scene, and it correctly interprets it and displays it.  It also shows multiple viewports on the same page.

JSON requires storing everything as strings, so it's far less than ideal for authoring data.  As an intermediate format it is probably as good as it's going to get.  A post-authoring conversion step might be in order.

Added two packages, class-transformer and reflect-metadata, to accomplish relatively hassle-free type-correct interpretation of JSON data.  This required annotating all class data members (no interfaces allowed) with decorators describing the type.

Disabled the max-classes-per-file 1 limitation as it is too limiting and unreasonable for these purposes.

WebGLObjectData no longer extends from Plane, Sphere and Cube data, as it is no longer an interface and can't extend from more than one class.  Instead if inheritance, it uses composition and has optionally one of each data member.