WebGL JSON from a URL fetch in posts list
Added hdrURL optional param to post data structure.  Right now it only handles a URL to WebGL JSON data.  It handles fully-qualified URLs or partial ones.  If not fully-qualified, it prepends the current window location host.  The content of the URL is loaded and parsed as JSON, then built as a WebGLScene and loaded into a WebGLViewport.

Currently it is only supported in the Posts list, and error handling is present but ugly and uninformative.  This is a save point checkin before making potentially breaking changes.

Renamed webglelement.ts to webglviewport.ts, because that's the name of the class now.

A missing visual or post now also says "this space unintentionally left blank" before the [ missing (whichever) ].

WebGL now stores an error string when it fails to find a canvas so that in the future widgets will be able to show a useful message.