WebGL JSON URL responsibility now in WebGLViewport, works on post items
Moved the WebGL JSON URL fetch code into the WebGLViewport class, which changed how these objects are instantiated in post_list and post_item.  Now they are still constructed on demand but without the scene data.  The scene data is passed in on init() so that it can be lazy-loaded.

Now fetchWebGLData, which lives in WebGLViewport, passes loaded scene data into init() on the same object when it's ready.

This magically makes URL-fetched WebGL JSON work for post_item as well.