<web-gl> to load, display WebGL content & fallbacks, SVG text wrapping
Added new WebGLElement component in webglelement.ts, which is embedded as <web-gl>.  It takes width, height, fontsize, and href -- a URL for JSON data.  It can handle width and height in percentage or pixel units.

Added a function to wrap SVG text, assuming it's monospaced (instead of embedding libraries to calculate rendered font sizes offscreen and other such nonsense).  It works well enough.

Added two SVG embeds to the WebGLElement in case the WebGL Scene isn't available: if there's an error, it shows the error text (wrapped) in light grey on a grey background.  This includes if WebGL isn't supported.  If there's no error but the webGL data isn't there (say, it hasn't loaded yet), an animated grey ellipsis will render instead.  This could probably be an animated PNG for cheaper.

Post_item now async loads the webglelement.ts module so it can show webgl content.  It also renders the post body content as unsafeHTML.  The post description has been moved above the divider.

Added hdrInline and hdrHref members to PostData interface to offer means to present inline HTML or an WebGL JSON or SVG/image URL, respectively.  The hdrAlt member is still used for alt text, but other hdr* members have been removed as they are now superfluous.

Cleaned up the presentation of the error visual SVG.

PostTile and PostItem no longer contain a WebGLViewport member directly -- they don't need to manually fetch and init WebGL content, because the WebGLElement component does that itself on-demand based on src data.  There isn't really a data path right now to pass in structured data directly, only string JSON.  However, the WebGLElement has a member that takes a WebGLScene: setWebGLData( scene ).

Added two new VS Code tasks: rebuild for dev and dist, to force rebuild their respective outputs.

Next:

Make a WebGLComponent

Make an SVG for loading and an SVG for an error state

Make a promise each for having presentable webgl data, and having a webGL error state

Use the "until" directive to show these in priority order inside the WebGLComponent

Figure out how to set up the viewport automatically
	CURRENT
		The viewport lives as an optional object on the parent component
		It can be created if requested, allocated and passed in the root and element to be created and placed
		Then the content is fetched if necessary and filled in via init()
	NEW
- 		A <web-gl> component exists in the DOM, so it doesn't need any root or element
		It has an optional href property which will be fetcted when assigned

WHEN THIS HREF IS FILLED IN, need to do a fetch and fill up the contents
Alternately, allow a direct construction and init with a structure from code or wherever

EXPERIMENT:  made a post body with a <web-gl> component in it, see if it works in post-item

- Wrapping text:
	Keep a property? array of "wrapped text" lines (can be undefined)
		Do property arrays force a render update?
	Observe for resize event
	On resize, re-wrap lines to force a render update (or do it manually)
	Get away with out ResizeObserver?
		Just do the work after firstUpdated and on document resize?