<web-gl> cheap fade-in when loaded, post info layout tweaks
Tweaked layout of post item.  Description sits better, image always sits at the right width and height .

In post tile, svg wasn't filling, but it does now.  Added CSS classes for post author and description (renaming previous description to info).  Tweaked layout.

Added cheap fade-in to <web-gl> once objects and all textures are loaded.  Right now it's just a pure alpha out fade, so alpha sort issues show through.  The fade is very fast so you don't really notice, though.  The better version of an opaque render to an offscreen texture and then a render with alpha would be nice but it's a a "would be nice".

Added notes and cml files in notes/ folder.