Markdown->HTML runtime post body conversion
Added in use of the Showdown md->HTML converter, which is loaded via <head> script fetch instead of as a module after too much trouble with module definitions.  Eventually, this could be gated on admin permission checks.

Now it takes the post body from data, which is assumed to be in Markdown format, and converts it to HTML.  Eventually this should already be in HTML format once the backend is set up properly.

Added some test post markdown body data.

Turned on esModuleInterop in tsConfig to play nice with Showdown md->HTML converter.