Admin page with basic Markdown->HTML editor
Added a rudimentary "admin" page, which is empty except for an "Editor" button.  That button pops up a page with two panels, left and right.  If you type Markdown on the left, it translates it to HTML on the right.

You can access the admin page directly from the URL only: /admin.

Added support for the MD->HTML conversion to set a class for each HTML tag it generates.

Updated my example first post after running over it in the editor.