New Admin embed: "List" showing the post list, allowing click or k/b select to open in Editor, listener cleanups
Added new editlist.ts element that embeds into Admin and acts as the Admin home screen.  When going there, you get a list of the defined posts.  You can scroll through with mouse or keyboard and click or hit Enter on one to open it in the Editor.  If you hit the Editor button, you'll get the blank Editor screen.

Lots of cleanup to listeners: all of the event listeners should now be cleaned up properly.

Editor loading sequence shuffled a bit to accomadate loading a post in by ID.  It now initializes the body MD text and does a one-time conversion at load time.  Soon it won't do this, probably, because the HTML will be cached, I suspect.

Tweaked coloring for navbars.