Navbar fixes/tweaks, data less static, header images more unified, fixed pagenav bugs, nicer hover effects
Fixed bug with navbar losing track of location.  I was removing the pageNav listener on leave, but it was only added in the constructor.  That wasn't a reliable pattern.  Instead, properly cleaned up all the listeners.  They should be added in "connectedCallback" and removed in "disconnectedCallback" if possible.  However, the DOM objects aren't all added yet in connectedCallback, so you may have to add in firstUpdated and remove in disconnectedCallback.  I do this in one place.

The only case that wasn't cleanly updated was WebGL resize/scroll listeners because it's a shitshow.  I noted the bug in my CML file.  Essentially, right now, it sets up new global window listeners for resize and scroll for each new viewport.  It only cleans up one each on navigate away.  If any resize or scroll event happens, it does a full render for each listener that fires, rendering ALL viewports no matter which was actually affected.  The ideal version has one listener set for each viewport, rendering only the affected ones on events.  The viewports hold the listeners, trigger renders, and clean them up themselves.  The WebGL object can still do global renders, but it seems like de-centralizing would be a win here.

Started the process of making my demo data less static.  There it's no longer const and has accessor functions.

Tags are now a comma-delimited string in source data.  The app is responsible for splitting into an array to work on it.  This way there are zero non-primitive objects in the data, making cloning trivial and predictable.

Unified the shape of header images and centralized the sizes in variables.  Cleaned up the various usage locations to handle img, SVG, and WebGL cases properly.

Added a pageNav event to the post_item page so that it updates the navbar correctly.  Before, if you navigate there directly or refresh, you'd get a navbar with no location.

Refined styling on navbars.

Added nice transitions on "Read More" and editor list table hovers.