WebGL elements are now obscured when scrolling as expected, cleaner <web-gl> loading/fade-in experience
Fixed <web-gl> elements showing on top even when they should be hidden due to scrolling.

Now the SVG loading animation dots fade in with a slight delay so that the WebGL has a chance to load first, looking like a cleaner experience for faster connections but a proper loading notification for slower ones.  The SVG loading anim also removes itself once the fade is complete.

Tidied some post data to be more immediately useful for testing.