WebGL component resize and scroll listeners working correctly, esp for static scenes, which also fade in
Instead of each viewport having listeners for resize and scroll (and not cleaning them up well), now only WebGL itself holds resize/scroll listeners (for <web-gl>) and it holds one set for the duration of the session.  It calls down to Elements to let them run their scroll logic.

Properly redrawing on scroll and resize, particularly when using non-animating WebGL scenes.

Fixed a bug with non-animated webgl scenes where they would seem to be blank grey the first time they render.  It turns out they were fading in but only showing the first frame of the fade-in.  Now when fading in they are treated as animating.

Made the base home page full-screen size.

Improved some of the debug render prints.

Added a second test WebGL scene, a static version of the first one.