Modified post tiles to Twind and new look (game device), new quips and style
Refactored layout for post tiles using Twind where possible (as @apply doesn't seem to work yet).  Now it's 2 wide still, but wider tiles filling the space.  The image portion has a border, the background is light grey, there's a drop shadow, and the whole thing resembles a handheld gaming device.  This is a good thing.

Slightly adjusted post list layout to be justify-evenly so that new form laid out nicely and it used new vars.

Added a few quips, and made them italic.