Fixed numerous WebGL scrolling, animating, refreshing event issues
Fixed pages not scrolling their contents back to top when navigating to a new page.  This was a change thanks to the use of a scrolling div instead of the window scroll.  Now, the listener is on the lit-content div, which uses the pageNav event to trigger a scroll to top.

This same change also triggered some modifications in how WebGL handles scrolling events.  lit-content now registers with WebGL to send it scrollevents.

WebGL no longer tries to transform the canvas by the scrollY.  Each viewport is rendering to the correct place on the screen-size canvas.

There is a new helper function isAnimating() to encapsulate the animated || isFading logic.

Renamed updateAnimated to updateAnimatedFlag for clarity.

WebGL tracks animated render request IDs and cancels ones where appropriate.  The scheme is:
	One render loop for all WebGL elements (if any)
	It only runs a loop if any are animating (ie, animated themselves, or fading in)
	If any animated, start animation loop
	If none animated, cancel animation loop
	If scroll or resize event, request new single render only if not animated

New functions: startAnimatedRender and stopAnimatedRender, taking and returning the request ID, to encapsulate and identify all places which do these things.

Refactored requestNewRender which started both animated and single renders into two functions:
- restartRendering()
	- this is the only place that can start the animated loop
	- called only during webglviewport::init and webgl::updateAnimatedFlag
- refreshSingleRender()
	- this checks first to see if it is animating (including fading in, so we don't double render during scroll), then queues a single render
	- called from resize, scroll, and navigateaway events (the latter after stopAnimatedRender is called)