Data: name, status, MD; hidden/visible posts, web-gl layout, saved posts
Updated post data to match desired DB schema: 'id' is now 'name'; new status field of 0 (invalid), 1 (hidden), or 2 (visible); 'body' has been replaced with two fields: 'markdown' and 'content' for the original MD and resulting HTML, respectively.

The status field is used to control whether a post is listed in the post_list.  It's reflected in the edit_list with a checkmark (which icon is now included in the assets folder) and editor with a toggle button.  When saving, if the status is hidden, it updates dateModified; if visible, it does not (as dateModified effectively becomes datePublished or dateUpdated at that point).

Now when a post is saved, it dumps the post's data in a copy/paste-ready format to the console.

Fixed the remaining slight border issue with web-gl that was my entirely caused by my border issue compensation fudge factor math.

Changed webglelement to make the host display: inline-block so that it lays out more predictably.  Moved it to styles.ts so it can be used in editor and everywhere else it's needed in code.

Also setting box-sizing: border-box and removing the special .webgl class style in post_item, now the <web-gl> element lays out the same in editor preview and post_item.  Setting inline-block on the web-gl-container also helped.

Turned off Twind class hashing for now.  Annoying that I can't get a dev/production variable into the code.

Tweaked a few quips and added a possible one.

Added a new post that's hidden.