Started <gl-code> element, post_item uses content data member instead of converting md, tweaked post display
There is a new <gl-code/> element that encapsulates a <split-panel/> with a <web-gl/> on the left and a <tab-group/> on the right, which currently shows 3 placeholder tabs (VS, FS, Tex).  The plan is for it to get an event on web-gl load and then grab the vs and fs source and perhaps the texture source as well, displaying them in the appropriate tabs for inspection and (hopefully) edification.

Made post_item use the post's content instead of re-running the markdown conversion.

Reformatted the post_item header display a bit.

Added the grip-vertical shoelace icon to the asset/icons folder.

Added *.lnk to .gitignore so I could store a shortcut to the shoelace icons folder in the asset icons folder.