PROD only: <gl-code> selecting objects, see shader code with syntax color, <web-gl> orbit/zoom, scrollbar tweaks
BROKE THE DEV BUILD.  Only the production build works with this checkin, because Prism steps require bundler functionality.  This checkin sets a functioning save point and the next will either transition not requiring that, or enabling dev bundling.

Added PrismJS to enabled syntax highlighting <pre><code> blocks.  Integration only partially worked and required adding babel-plugin-prismjs and @rollup/plugin-commonjs.  The latter may help with other UMD modules like aws-sdk.  In the end I got it sort of working by manually including the minified CSS in my Web Component and manually running the highlight step on the code, then including the unsafeHTML.  This disables any of the Prism plugins, however, some of which operate directly on the DOM.  I'm not sure if many/any of them are possible in a lit-type Web Component framework, but I have basic syntax highlighting now and it only runs when you change object.

I tuned the highlights to make various keywords more distinct.  Hopefully this will help people to recognize the terms more easily, like varying vs uniform.  They do need a colorblind safety pass, like the rest of the site.

Disabled rollup-plugin-import-meta-assets as it was causing problems.
Added rollup-plugin-styles.
Reactivated plugin-css-only.
Reactivated copying all *.css to dist.

Added in the babel prismjs step and configured it with my language, plugins, and theme, and instructed it to export css.  It seems to be necessary to have these steps in this configuration.

Added an in-between gray 350 color.

Added a class to "prevent select" for preventing text selection.  This is used to allow clicking and dragging for other element reasons.

Overhauled scrollbar appearance to be slicker and work with light and dark backgrounds, with better hover/active coloring.

Can now left click and drag to orbit the WebGL camera.

Can now right click and drag vertically to zoom WebGL camera.

Can now use mouse wheel to zoom WebGL camera.

GLCode element supports changing object selection with two arrows in the bottom bar.  When you do, it shows the selected vertex shader/fragment shader in the respective tabs.  It populates the texture URL in the 3rd, but that isn't showing correctly yet.

GLCode imports prismjs, the minified glsl language plugin, and the minified theme.  I pasted the calculated Prism CSS data inline in the styles element.

I got the scrolling tabs of shader code working as expected in the resizable split-tabs element, with max/min settings and a center snap.

There's a dummy FPS/ms counter in the lower left of the GLCode status bar.

Tweaked the FAQ content.

Added left/right filled caret icons.