Framerate calc and display
There is now an active framerate counter in webglviewport.  During each render, if a second has passed since the last fps recording, it takes a new measurement by comparing frames vs time delta and stores it, resetting the start time.  It also fires off a custom event to notify any listeners that a new, up-to-date fps for this element is available.

On each call to the "getAverageFrameDuration" method, if it's been more than a second since the benchmark was taken, it returns 0.0.  This way it handles reverting to unanimated.

This mp/fps is displayed in the <gl-code> status bar.

Made post header dark gray.