New dev web server, tweaked shader code and resize handle, fixed site name, <gl-code> auto-selects first, redirect on invalid URL
Added a new dev web server (serve, by vercel) to replace webdevserver, which the Prod build used but failed every second build.  It was caching index.html regardless of my .htaccess file, so I had to manually kill and restart it.  Unfortunately, serve holds onto file locks so you can't delete the dist/ folder and rebuild, so the new flow is for the prod build to build and run serve, and you restart the task each time.  It's a reasonable flow now, however, and reliable.  The serve.json config file is in one place and it would be trivial to configure a dev server to run on another port in parallel if desired.

Tweaked shader code output to minimize whitespace and comments at top and bottom.

Made index.html show "Subpixel" in conventional case, instead of camelCase.

Tweaked size of split-bar resize handle.

<gl-code> panel now selects first object on load.

If you load an invalid URL, you'll now get redirected to the home page via a Vaadin route fallback.