gl-code caching, shaded rendering, more post data, tweaked faq, improved web-gl scene and tech abilities
Moved Showdown to just Editor.  Added an extension to filter the converted the result and add an "id=<n>" field to all <gl-code> elements in the post content, in order.  This is consistent and stable during editing unless changing the count  of <gl-code> objects.  Then these are the IDs used to define the container div name, and the IDs used as the cache keys for viewports.  The full outerHTML is still used as a hash for the cached viewport to ensure validity.  If it changes, then the cache entry is removed and recreated.  Otherwise, the viewport is reconnected to the new element, set up to display, and returned.

This means that it is now possible to edit for prolonged periods with <web-gl> loaded without performance problems, and should be possible to load webgl at the start.

Added an experimental (and functional) extension that enforced two spaces after periods but disabled it when a copy/paste made it clear that unicode space values after each sentence could cause issues for people if they try to copy out any info.  It's not worth it.

Added shaded-box-spin.json to be the second populated gl-code box.  It also requires stretching the systems a bit: showing a lit object, and hence defining lights and more camera controls, including an orthographic option.

Modified unshaded-box-spin.json to be an orthographic unlit spinning cube from a deceptive upwards-looking angle so that the following lit draw from the same angle will be a surprise.

Renamed WebGL object anim "mode" and "modeIn" variables to "blend" and "blendIn" to be more clear.

Tweaked the FAQ some more.

Made <web-gl> scaling a bit more robust in various cases.

Added new shaders: col_diff.fs for color and diffuse lighting, and posnorm.vs for position and normal.

Edited the constants and comments on all shaders to be consistent and clean.

WebGLScene members added: orthoDiag (diagonal orthographic scene size), lightDeg (dir light yaw/pitch in degrees), lightColor (dir light color), ambientLight (ambient color), ambientStr (ambient strength).