Alert embed markdown, new supporting icons, faq update
Added <alert/> markdown syntax to embed an sl-alert field with a specified variant.  There's commented-out support for specifying a specific icon as well, but the 5 default with the default variants should do for now.

Added the 5 icons for the 5 default sl-alerts.

Tweaked the FAQ: uploaded an image of my markup language to-do list.