New dev rollup build, legal footer
Reduced npm build/run/test/analyze/rebuild/clean tasks to just 4: build/build&run test&prod.  The build and run for Test and Prod run web dev servers on 8010 and 8011 out of test/ and dist/, respectively, so they can run simultaneously.

Added @rollup/plugin-replace to support telling the project if it is running in production or development mode.  It also tells it the build date/time.  This information is revealed in the legal footer.

Added a legal footer to all public-facing pages that uses a copyright C circle icon and Chris Lambert 2023 along with the build type and date/time.

The new legal footer is hidden if requested, for example in the admin section.

Modified the rollup config to support a test rollup bundled build to test/ instead of dist/.  It turns out I have to use rollup to process node_module with commonJS to make the build work now, so I can't just run against my raw JS anymore, apparently.  Right now it's a rebuild every time, but that can be easily changed for testing and there's commented out code to also support a watch mode.

Removed the "Web Development" question from the FAQ after feedback.