Posts changed to Elements, elements loading from DB, gl xform improvements, content updates
Changed database from post_data to elements, and it now contains a list of elements that can be posts, tags, or threads.  They have a partition key of "name" and a sort key of "type" (an enum ElementType of Post, Tag, or Thread).  PostStatus is now ElementStatus because a thread could share it as well.

Tags are now implemented as an array of strings, which works better with DB query support.

Added "next" field to Element (formerly Post) structure, which will be used for threads eventually.

The local posts list is now the devPosts list, and only filled in on "development" builds.  This way there is extra content for test builds but only DB content for any production build.

Post_item reports an error with a clean-ish message when it fails to get the post from the DB.

The editor tries to commit to the dialog on "Save".  Currently it fails due to a lack of proper authentication for the authenticated AWS role.

The editor puts up a modal dialog on any db put failure.

Added DynamoDBDocumentClient support for supposedly easier database access, though primary read and write functions still use raw DBClient access.

Added util-dynamodb package to get access to marshall and unmarshall and NativeAttributeValue support to use with the DynamoDB non-DocumentClient access technique.

Removed commented items from rollup.config.js to improve readability (no functionality changes).

Removed web development question link from FAQ as question was removed previously.  Updated rest of FAQ slightly.

Fixed <ul> style to match the rest of the post.

Added 'rootxform' element to WebGLData to be used as a base transform on object, applied before the next (single) layer of potentially-animated xform elements.

Now normalizing rotAxis in data transform objects.

You can now scale the ortho diagonal up and down with the middle mouse button the same way you scale the perspective FOV up and down with the middle mouse button (via scroll) or via L+R click and move up/down.

Added shaded-sphere-spin json for basicshading post, which is now on the DB.  It shows a minimally-tesselated sphere next to a cube rotated as a diamond also spinning, so that they compare independent-normal 'flat' shading and shared-normal 'gouraud-interpolated' shading.

Updated the poscol-cube json to actually show a cube with the position-as-color shaders.